requirements.txt
celery==5.4.0
Django==4.2.6
django-celery-beat==2.6.0
django-ckeditor==6.7.1
django-cors-headers==4.3.0
django-countries==7.5.1
django-filter==23.3
djangorestframework==3.14.0
Pillow==10.1.0
psycopg2==2.9.9
redis==5.0.4
stripe==15.0.1
